Changed floats to doubles to be consistent

+++ src/Domain/Map/MapNode.java
@@ -15,9 +15,9 @@ public class MapNode
     int posY;
     int nodeID;
     UID nodeUID;
-    float g = 0;
-    float heuristic = Float.MAX_VALUE;
-    float f = Float.MAX_VALUE;
+    double g = 0;
+    double heuristic = Double.MAX_VALUE;
+    double f = Double.MAX_VALUE;
     NodeEdge parent = null;
     Image node = null;
     Floor myFloor;
@@ -26,38 +26,26 @@ public class MapNode
     public int getPosX() {
         return posX;
     }
-
     public int getPosY() {
         return posY;
     }
-
-    public float getG() {
+    public double getG() {
         return g;
     }
-
-    public float getHeuristic() {
+    public double getHeuristic() {
         return heuristic;
     }
-
-    public float getF() {
+    public double getF() {
         return f;
     }
-
     public NodeEdge getParent() {
         return parent;
     }
-
-    public HashSet<NodeEdge> getEdges() {
-        return edges;
-    }
-
-    public void setG(float g) {
+    public HashSet<NodeEdge> getEdges() {return edges;}
+    public void setG(double g) {
         this.g = g;
     }
-
-    public void setFloor(Floor f) {
-        this.myFloor = f;
-    }
+    public void setFloor(Floor f) {this.myFloor = f;}
 
     public boolean hasEdgeTo(MapNode n) {
         for (NodeEdge e: edges) {
@@ -68,11 +56,11 @@ public class MapNode
         return false;
     }
 
-    public void setHeuristic(float heuristic) {
+    public void setHeuristic(double heuristic) {
         this.heuristic = heuristic;
     }
 
-    public void setF(float f) {
+    public void setF(double f) {
         this.f = f;
     }
 
@@ -114,8 +102,8 @@ public class MapNode
 
     public void resetTempValues() {
         this.g = 0;
-        this.heuristic = Float.MAX_VALUE;
-        this.f = Float.MAX_VALUE;
+        this.heuristic = Double.MAX_VALUE;
+        this.f = Double.MAX_VALUE;
         this.parent = null;
     }
